+++ DKEngine/MarIO/Assets/Models/Enemy.cs
@@ -1,6 +1,4 @@
 ﻿using DKEngine.Core;
-using DKEngine.Core.Components;
-using MarIO.Assets.Scripts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,7 +9,7 @@ namespace MarIO.Assets.Models
 {
     class Enemy : GameObject
     {
-        public enum EnemyType
+        enum EnemyType
         {
             Goomba,
             GoombaBlue,
@@ -31,7 +29,7 @@ namespace MarIO.Assets.Models
 
         Dictionary<EnemyType, string> EnemyTypeNames = new Dictionary<EnemyType, string>()
         {
-            { EnemyType.Goomba, "goomba" },
+            { EnemyType.Goomba, "" },
             { EnemyType.GoombaBlue, "" },
             { EnemyType.GoombaSilver, "" },
             { EnemyType.KoopaTroopa, "" },
@@ -47,8 +45,6 @@ namespace MarIO.Assets.Models
             { EnemyType.BillBlasterSmall, "" }
         };
 
-        public EnemyType Type { get; set; }
-
         public Enemy()
             :base()
         { }
@@ -60,47 +56,6 @@ namespace MarIO.Assets.Models
         protected override void Init()
         {
             this.Name = "Enemy";
-            //this.TypeName = EnemyTypeNames[Type];
-
-            switch (Type)
-            {
-                case EnemyType.Goomba:
-                    this.InitNewComponent<Collider>();
-                    this.Collider.Area = new System.Drawing.RectangleF(0, 0, 16, 16);
-
-                    this.InitNewScript<GoombaController>();
-                    this.InitNewComponent<Animator>();
-                    this.Animator.AddAnimation("default", Database.GetGameObjectMaterial(EnemyTypeNames[Type]));
-                    break;
-                case EnemyType.GoombaBlue:
-                    break;
-                case EnemyType.GoombaSilver:
-                    break;
-                case EnemyType.KoopaTroopa:
-                    break;
-                case EnemyType.KoopaParatroopa:
-                    break;
-                case EnemyType.PiranhaPlant:
-                    break;
-                case EnemyType.Spiny:
-                    break;
-                case EnemyType.BuzzyBeatle:
-                    break;
-                case EnemyType.BuzzyBeatleBlue:
-                    break;
-                case EnemyType.BuzzyBeatleSilver:
-                    break;
-                case EnemyType.FireBar:
-                    break;
-                case EnemyType.BulletBill:
-                    break;
-                case EnemyType.BillBlasterLarge:
-                    break;
-                case EnemyType.BillBlasterSmall:
-                    break;
-                default:
-                    break;
-            }
         }
     }
 }
