+++ DKEngine/MarIO/Assets/Models/Mario.cs
@@ -1,6 +1,5 @@
 ﻿using DKEngine.Core;
 using DKEngine.Core.Components;
-using DKEngine.Core.UI;
 using MarIO.Assets.Scripts;
 using System;
 using System.Collections.Generic;
@@ -12,34 +11,11 @@ namespace MarIO.Assets.Models
 {
     class Mario : GameObject
     {
-        public enum State
-        {
-            Small,
-            Super,
-            Fire,
-            Invincible
-        };
+        public CharacterController.State State;
 
         protected override void Init()
         {
             this.Name = "Player";
-
-            Trigger left = new Trigger(this);
-            left.Name = "Left_Trigger";
-            left.InitNewScript<LeftMarioChecker>();
-
-            Trigger right = new Trigger(this);
-            right.Name = "Right_Trigger";
-            right.InitNewScript<RightMarioChecker>();
-
-            Trigger top = new Trigger(this);
-            top.Name = "Top_Trigger";
-            top.InitNewScript<TopMarioChecker>();
-
-            Trigger bottom = new Trigger(this);
-            bottom.Name = "Bottom_Trigger";
-            bottom.InitNewScript<BottomMarioChecker>();
-
             //this.TypeName = "mario";
             this.Transform.Position = new Vector3(50, -10, 0);
             this.InitNewComponent<Animator>();
