+++ DKEngine/MarIO/Assets/Scripts/BonusBlockScript.cs
@@ -1,0 +1,27 @@
+﻿using DKEngine.Core;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using DKEngine.Core.Components;
+
+namespace MarIO.Assets.Scripts
+{
+    class BonusBlockScript : Script
+    {
+
+
+        public BonusBlockScript(GameObject Parent) : base(Parent)
+        { }
+
+        protected override void OnColliderEnter(Collider e)
+        { }
+
+        protected override void Start()
+        { }
+
+        protected override void Update()
+        { }
+    }
+}
