+++ DKEngine/MarIO/MarIO.csproj
@@ -53,17 +53,14 @@
     <Compile Include="Assets\Models\Enemy.cs" />
     <Compile Include="Assets\Models\Group.cs" />
     <Compile Include="Assets\Models\Mario.cs" />
-    <Compile Include="Assets\Models\Trigger.cs" />
     <Compile Include="Assets\Scenes\Test.cs" />
-    <Compile Include="Assets\Scripts\EnemyControllerScripts.cs" />
-    <Compile Include="Assets\Scripts\MarioTriggerColliderScripts.cs" />
+    <Compile Include="Assets\Scripts\BonusBlockScript.cs" />
     <Compile Include="Assets\Scripts\CameraController.cs" />
     <Compile Include="Assets\Scripts\CharacterController.cs" />
     <Compile Include="Assets\Scripts\PipePort.cs" />
     <Compile Include="Assets\Scripts\PlayerControl.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Shared.cs" />
     <Compile Include="Sprites\Enemies.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
